Enhancement: Generic Address Space Checking
In this commit, I introduce some rework in the mm_is_uaddr() and
mm_is_kaddr() helper functions, so that the virtual memory layout of
Nanvix can be more flexible. What we assume is that:

(a) The kernel has two big regions - the kernel code data and the kernel
page pool, which not necessarily contiguous;

(b) The user space is a single and contiguous memory area.