Bug Fix: Bad Invalid Free Frame Test
In this commit, I fix the faulting injection test for freeing an invalid
page frame. We were supplying the page frame id instead of the page
frame number, and thus the test was failing with the bad error code.
Notice that I have also moved the previous test to the right  unit test
(bad free).