Bug Fix: Avoid Linking LibGCC with the Kernel
In this commit, I introduce a workaround in the kvsprintf() function in
order to avoid the linkage of LibGCC routines to compute integer
divisions. Machine code of these functions seems to be buggy in the k1b
architecture, causing invalid opcode exceptions to arise in kernel land.

While these workaround does remove the decimal formatted print output
from the kernel, it is important to note that we do not **really** need
rich output in kernel land.