[k1b] Feature: Virtual Exceptions
In this commit, I introduce support for virtualised exceptions in the
k1b architecture. Some exceptions are not directly triggered by the
hardware, but are raised by other exceptions. For instance, a TLB Fault
may incur in a Page Fault, if no virtual-to-physical address mapping is
currently present in software-maintained page tables.