[hal] Enhancement: Splitting TLB Miss into Data and Instruction
Some architectures have a distinct TLB for data and instruction
respectively. In these archs, is very important to distinguish
between both, because not doing so will leads to wrong writes
int TLB, wrong permissions and even starving (if a data miss
was served as instruction miss).

In order to deal with, this commit splits the the misses in
instructions and data. Although the handler itself is the same,
the miss type can be checked as the exception number in the
exception structure.

Besides that, the tlb_write() was removed from the upage_map(),
since this could lead to wrong mapping in architectures that use
direct mapping, since the current mapping could be overwritted
for a different mapping that would be used in the future.