[k1b] Bug Fix: Races When Putting Core to Sleep
In this commit, I introduce a fix for a race condition that might happen
when a core was put to sleep. Overall, the startup signal could be lost
and if a core received multiple signals it might loose some of them. To
fix this, I have redesigned the state transition between cores and I
have introduced a counter for the number of received wakeup signals.