[k1b] Bug Fix: Bad Context Structure Layout
In this  commit, I have fixed two bugs:

- The layout of the context structure in the k1b architecture
was wrong (sizes of ps and sps registers are 8 bytes). In this commit, I
have fixed this bug.

- We were reading the values of spc, ps and sps registers directly from
hardware. Instead, we should read these from the mOS scoreboard.

Furthermore, I have changed the do_generic_excp() routine so that to
dump correctly the machine state.