[or1k] Bug Fix: Initial TLB Handler
* [or1k] Bug Fix: Initial TLB Handler

According to the issue #91, the OpenRISC architecture needs a TLB miss
handler in the very early stages, since a kernel miss can occur during
the kernel tests.

Thus, this commit introduces a private TLB handler that is set in the
core setup and it's used until the HAL TLB handler is setup in the
upool_init().

* [or1k] Enhancement: No Initial Page Fault Handler

In this commit, I have removed the initial page fault handler. We
hopefully do not trigger page faults when starting up the kernel. This
is likely to be an error.