Enhancement: Check for Bad Thread Start Routine
In this commit, I have introduced checks in the kthread_create() kernel
call for bad thread start routine addresses. Note that this feature is
not turned on by default, because in some targets the kernel may be
linked with the user binary, thus resulting in error.

Enhancement: Check for Bad Thread Argument

In this commit, I have introduced checks in the kthread_create() kernel
call for bad thread argument address. Note that this feature is not
turned on by default, because in some targets the kernel may be linked
with the user binary, thus resulting in error.

Enhancement: Check for Bad Thread Join

In this commit, I have introduced checks in the kthread_join() kernel
call for bad thread return value address. Note that this feature is not
turned on by default, because in some targets the kernel may be linked
with the user binary, thus resulting in error.

Enhancement: Check for Bad Thread Exit

In this commit, I have introduced checks in the kthread_exit() kernel
call for bad return value address. Note that this feature is not turned
on by default, because in some targets the kernel may be linked with the
user binary, thus resulting in error.