Squashed commit of the following:
Enhancement: Removing Debugging Output

Bug Fix: Signal Thread Exit and Wait

Previously, we were not waiting for a thread that has terminated to be
joined before ripped out. In this commit, I have fixed this bad
semantic. Now, when a thread terminates, it waits for on a condition
variable before releasing underlying resources.