Bug Fix: All Exiting Threads Unblock
Previously, we had a single conditional variable for signalling that a
joining thread was ripped out a thread that was exiting. Thus, if more
than one thread was exiting, all of them would be unblocked, even if
no thread was joining them. In this commit, I have fixed this problem.