[k1b] Enhancement: Save Stack Registers
In this commit, I introduce a change in the context information
structure of the k1b architecture, so that stack registers (ie SSP, SSSP
and SSSSP) are saved when an exception occurs. Additionally, I
have hacked the test driver for the HAL so that we can test this new
feature.