[pm] Feature: Kernel Events.
In this commit, I introduce the kernel events interface.
The kernel events rely on the HAL-level events. Specifically, we provide
an event handler that is always called when an event arrives on the
target core. The HAL defines if this event will be an IPI interrupt or
an explicit-waiting event.

Over this interface, we seek to implement preemptive threads on
architectures that support preemptive events (IPI).

New files:
	include/nanvix/kernel/event.h
	src/kernel/pm/event.c